chore: move session recording query tests from EE to OSS
