+++ posthog/session_recordings/persist_to_lts/__init__.py
@@ -1,0 +1,9 @@
+from posthog.session_recordings.persist_to_lts.persistence_tasks import (
+    persist_finished_recordings_v2,
+    persist_single_recording_v2,
+)
+
+__all__ = [
+    "persist_finished_recordings_v2",
+    "persist_single_recording_v2",
+]
