+++ posthog/session_recordings/persist_to_lts/session_recording_extensions.py
