+++ posthog/session_recordings/playlist_counters/test/test_recordings_that_match_playlist_filters.py
@@ -22,13 +22,12 @@ from posthog.redis import get_client
 from posthog.session_recordings.models.session_recording import SessionRecording
 from posthog.session_recordings.models.session_recording_playlist import SessionRecordingPlaylist
 from posthog.session_recordings.models.session_recording_playlist_item import SessionRecordingPlaylistItem
-from posthog.session_recordings.session_recording_playlist_api import PLAYLIST_COUNT_REDIS_PREFIX
-
-from ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters import (
+from posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters import (
     DEFAULT_RECORDING_FILTERS,
     count_recordings_that_match_playlist_filters,
     enqueue_recordings_that_match_playlist_filters,
 )
+from posthog.session_recordings.session_recording_playlist_api import PLAYLIST_COUNT_REDIS_PREFIX
 
 
 class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
@@ -43,7 +42,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         mock_capture_exception.assert_not_called()
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_that_match_no_recordings(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -67,7 +68,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         }
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_that_match_recordings(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -100,7 +103,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         }
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_that_match_recordings_records_previous_ids(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -136,7 +141,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         }
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_that_match_recordings_skips_cooldown(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -158,7 +165,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         assert self._get_counts_from_redis(playlist) == existing_value
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_matching_legacy_filters(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -233,7 +242,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         )
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_skips_default_filters(self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock):
         playlist = SessionRecordingPlaylist.objects.create(
             team=self.team,
@@ -246,9 +257,11 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
 
     @snapshot_postgres_queries
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
     @patch(
-        "ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
     )
     def test_sorts_nulls_first_and_then_least_recently_counted(
         self, mock_count_task: MagicMock, _mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
@@ -292,7 +305,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
             ]
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_template_rageclick_filter_should_process(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ) -> None:
@@ -353,7 +368,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         )
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_with_too_many_errors_skips(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -375,7 +392,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         assert self._get_counts_from_redis(playlist) == existing_value
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_with_too_recent_error_skips(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -397,7 +416,9 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         assert self._get_counts_from_redis(playlist) == existing_value
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
     def test_count_recordings_only_queries_since_last_count(
         self, mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
     ):
@@ -458,9 +479,11 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         return json.loads(counts.decode("utf-8"))
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
     @patch(
-        "ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
     )
     def test_excludes_default_template_playlists(
         self, mock_count_task: MagicMock, _mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
@@ -478,9 +501,11 @@ class TestRecordingsThatMatchPlaylistFilters(APIBaseTest, QueryMatchingTest):
         assert playlist.id not in [call_args[0][0] for call_args in mock_count_task.delay.call_args_list]
 
     @patch("posthoganalytics.capture_exception")
-    @patch("ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query")
     @patch(
-        "ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.list_recordings_from_query"
+    )
+    @patch(
+        "posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters.count_recordings_that_match_playlist_filters"
     )
     def test_excludes_playlists_with_pinned_items(
         self, mock_count_task: MagicMock, _mock_list_recordings_from_query: MagicMock, mock_capture_exception: MagicMock
