+++ posthog/tasks/tasks.py
@@ -7,7 +7,6 @@ from django.db import connection
 from django.utils import timezone
 
 import requests
-import posthoganalytics
 from celery import shared_task
 from prometheus_client import Counter, Gauge
 from redis import Redis
@@ -808,22 +807,16 @@ def check_flags_to_rollback() -> None:
 
 @shared_task(ignore_result=True)
 def ee_persist_single_recording_v2(id: str, team_id: int) -> None:
-    try:
-        from ee.session_recordings.persistence_tasks import persist_single_recording_v2
+    from posthog.session_recordings.persist_to_lts.persistence_tasks import persist_single_recording_v2
 
-        persist_single_recording_v2(id, team_id)
-    except ImportError:
-        pass
+    persist_single_recording_v2(id, team_id)
 
 
 @shared_task(ignore_result=True)
 def ee_persist_finished_recordings_v2() -> None:
-    try:
-        from ee.session_recordings.persistence_tasks import persist_finished_recordings_v2
-    except ImportError:
-        pass
-    else:
-        persist_finished_recordings_v2()
+    from posthog.session_recordings.persist_to_lts.persistence_tasks import persist_finished_recordings_v2
+
+    persist_finished_recordings_v2()
 
 
 @shared_task(
@@ -831,15 +824,11 @@ def ee_persist_finished_recordings_v2() -> None:
     queue=CeleryQueue.SESSION_REPLAY_GENERAL.value,
 )
 def count_items_in_playlists() -> None:
-    try:
-        from ee.session_recordings.playlist_counters.recordings_that_match_playlist_filters import (
-            enqueue_recordings_that_match_playlist_filters,
-        )
-    except ImportError as ie:
-        posthoganalytics.capture_exception(ie, properties={"posthog_feature": "session_replay_playlist_counters"})
-        logger.exception("Failed to import task to count items in playlists", error=ie)
-    else:
-        enqueue_recordings_that_match_playlist_filters()
+    from posthog.session_recordings.playlist_counters.recordings_that_match_playlist_filters import (
+        enqueue_recordings_that_match_playlist_filters,
+    )
+
+    enqueue_recordings_that_match_playlist_filters()
 
 
 @shared_task(ignore_result=True)
